cli: move client import inside functions
* Client operations are expensive, moving the client
  imports inside the functions that use them reduces
  loading time as it only performs the import when it's
  really needed
